+++ src/components/tree-item/UTreeItem.ts
@@ -116,7 +116,7 @@ export class UTreeItem extends UElement {
           @click=${this.handleCheckboxClick}
         >
           <u-icon lib="internal"
-            name=${this.indeterminate ? 'dash-lg' : 'check-lg'}
+            name=${this.indeterminate ? 'minus' : 'check'}
           ></u-icon>
         </span>