+++ src/index.ts
@@ -16,6 +16,7 @@ export * from './components/card/UCard.js';
 export * from './components/carousel/UCarousel.js';
 export * from './components/checkbox/UCheckbox.js';
 export * from './components/chip/UChip.js';
+export * from './components/copy-button/UCopyButton.js';
 export * from './components/dialog/UDialog.js';
 export * from './components/divider/UDivider.js';
 export * from './components/drawer/UDrawer.js';